refactor(mpc): Separate between type and constant value exports
