+++ packages/mpc/src/constants.ts
@@ -1,0 +1,11 @@
+
+export enum StoragePlatform {
+	Browser = "browser",
+	CLI = "cli",
+}
+
+export enum WalletId {
+	MetaMask = "metamask",
+	Stackup = "stackup",
+	Biconomy = "biconomy",
+}
